refactor(neo4j): move display module to /examples
