+++ rig-neo4j/examples/vector_search_movies_add_embeddings.rs
@@ -19,6 +19,9 @@ use rig_neo4j::{
 };
 use serde::{Deserialize, Serialize};
 
+#[path = "./display/lib.rs"]
+mod display;
+
 #[derive(Debug, Deserialize, Serialize, Clone)]
 struct Movie {
     title: String,
@@ -116,20 +119,15 @@ async fn main() -> Result<(), anyhow::Error> {
         .top_n::<Movie>("a historical movie on quebec", 5)
         .await?
         .into_iter()
-        .map(
-            |(score, id, doc)| rig_neo4j::vector_index::display::SearchResult {
-                title: doc.title,
-                id,
-                description: doc.plot,
-                score,
-            },
-        )
+        .map(|(score, id, doc)| display::SearchResult {
+            title: doc.title,
+            id,
+            description: doc.plot,
+            score,
+        })
         .collect::<Vec<_>>();
 
-    println!(
-        "{:#}",
-        rig_neo4j::vector_index::display::SearchResults(&results)
-    );
+    println!("{:#}", display::SearchResults(&results));
 
     let id_results = index
         .top_n_ids("What is a linglingdong?", 1)
